+++ src/base/components/guard/RenderGuard.tsx
@@ -1,0 +1,12 @@
+/*
+ * Copyright (C) Contributors to the Suwayomi project
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+import type { PropsWithChildren } from 'react';
+
+export const RenderGuard = ({ condition, children }: PropsWithChildren<{ condition: boolean }>) =>
+    condition ? children : null;