+++ src/base/components/guard/ShareGuard.tsx
@@ -7,11 +7,8 @@
  */
 
 import type { PropsWithChildren } from 'react';
+import { RenderGuard } from '@/base/components/guard/RenderGuard.tsx';
 
-export const ShareGuard = ({ children }: PropsWithChildren) => {
-    if (!navigator.share) {
-        return null;
-    }
-
-    return children;
-};
+export const ShareGuard = ({ children }: PropsWithChildren) => (
+    <RenderGuard condition={!!navigator.share}>{children}</RenderGuard>
+);